// src/app/shared/model/periodeInventaire.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PeriodeInventaire {
    id_periode_inventaire?: number;
    id?: number;
    libelle?: string;
    n_bordereau?: string;
    isActive?: boolean;
    date_debut?: string;
    date_fin?: string;
}